+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](semver.md).
 ### Added
 - `string` extension for Random
 - Directory as Directory2 alias
+- `luhn10` extension for String
 ### Changed
 ### Deprecated
 - Queue (`klib.queue.Queue`)